v2.2.12 update push
Switch to distribute textures as an assetbundle and leverage KSP Texture Loader as a dependency. Keeping the textures within the OPM/OPM_Textures/PluginData folder structure within this repository, for posterity, and not committing the assetbundle as it requires Git LFS. It can be found within the distributed release file.